Add Benchmarking to main branch
In order to add benchmarks to the scrape route some minimal refactoring
will be required. It will also require the ability to create mocks for
the commands to create a fake response that can be parsed. We will also
include the github action that will run the benchmark, wrapped in
bencher. This will allow us to get a starting point to compare against
the prometheus client crate PR.

Signed-off-by: William Johnson <wjohnson@whamcloud.com>